+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Renamed *beat at start* to *beat in  bar* in beat marker
 - Split track/marker colors into an RGB color code and a color index
+- Add optional sample position to marker positions
 
 ## [0.7.2] - 2020-05-03